1. Commands "ip-admin-add" and "ip-admin-remove" were renamed to "admin-add" and "admin-remove". 2. Re-implemented "admin-add" and "admin-remove" procedures, now they work properly, with any IP/IP-range. 3. Added new command "admin-list" that lists all admin IPs and admin CIDRs from OpenStack. 4. Added new command "execute-bastion>>>" that executes a command or a simple set of commands (Bash) on Bastion. 5. Added new command "execute-master>>>" that executes a command or a simple set of commands (Bash) on Master. 6. Added tool launcher that should work on (K)Ubuntu and CentOS - double-click "Launcher.desktop" to launch the tool. It will promtp for OpenStack username/password once and remember them (stored in temp/login). 7. Added switchable auto-deletion of older logs (set "maxLogs" in config). 8. JClouds libs updated to v2.0.1. 9. Fixed bug "SLF4J: Failed to load class org.slf4j.impl.StaticLoggerBinder. Defaulting to no-operation (NOP) logger implementation". 10. Changed help-text representation and cleaned up the output printed when the tool starts. 11. Various small improvements/fixes. 12. META-pipe version updated to v40/v13. 13. Updated documentation.

+++ v4 (dev)/out/sw_files/_prepare.sh
@@ -14,14 +14,18 @@ if [ "$1" == "sw-artifacts-prepare" ]; then
     ARTIFACTS_USERNAME="$2"
     ARTIFACTS_PASSWORD="$3"
     cd /media/$SW_MAIN_DIR
-    sudo mkdir $SW_FILES_DIR_NAME
+    sudo mkdir -p $SW_FILES_DIR_NAME
     cd $SW_FILES_DIR_NAME
     sudo chmod 777 .
     curl -O -u $ARTIFACTS_USERNAME:$ARTIFACTS_PASSWORD $ARTIFACTS_FILES_EXEC
-    curl -u $ARTIFACTS_USERNAME:$ARTIFACTS_PASSWORD $ARTIFACTS_FILES_DEPS_ARC | tar xv -C .
-    # Alternative: tar xvz< <(curl -u $ARTIFACTS_USERNAME:$ARTIFACTS_PASSWORD $ARTIFACTS_FILES_DEPS_ARC)
-    sudo mv package/dist ./metapipe
-    sudo rmdir package
+    if [ ! -d "metapipe" ]; then
+        curl -u $ARTIFACTS_USERNAME:$ARTIFACTS_PASSWORD $ARTIFACTS_FILES_DEPS_ARC | tar xv -C .
+        # Alternative: tar xvz< <(curl -u $ARTIFACTS_USERNAME:$ARTIFACTS_PASSWORD $ARTIFACTS_FILES_DEPS_ARC)
+        sudo mv package/dist ./metapipe
+        sudo rmdir package
+    else
+        echo "Deps directory 'metapipe' exists, skipping downloading/unpacking of deps."
+    fi
     echo "{
       \"metapipeHome\": \"/$SW_PARENT_DIR/$SW_MAIN_DIR/$SW_FILES_DIR_NAME/metapipe\",
       \"metapipeTemp\": \"/$SW_PARENT_DIR/$SW_TMP_DIR/metapipe-tmp\"
